Fix vaultTransfer wire format and improve error surfacing for malformed requests
Running the examples against testnet surfaced two real bugs:

- vaultTransfer's `usd` field was sent as a literal dollar integer, but
  the wire format wants raw USDC units (x1e6). Requests for e.g. "$5"
  were actually being interpreted as $0.000005, so every deposit was
  silently rejected as below the $5 minimum. VaultTransferRequest::usd
  is now a double (dollars, matching every other amount field in the
  SDK - UsdSendRequest, SpotSendRequest, etc.), and
  ExchangeRequestBuilder::vaultTransfer does the x1e6 scaling
  internally so callers don't have to.

- rest_vault_transfer.cpp's example vaultAddress was missing a trailing
  hex digit (39 chars instead of 40), which the server rejected at the
  JSON-body deserialization layer with a plain-text (non-JSON) error.
  parseSimpleResponse now falls back to surfacing that raw text as
  response.error instead of silently returning a blank status/error
  when the whole body isn't valid JSON.

Verified live on testnet: vaultTransfer now correctly reaches
vault-specific business logic (rejected for an unrelated reason -
"Deposit causes leader share to drop below 5%" - rather than failing
at the transport layer).

+++ include/hyperliquid/types/RequestTypes.h
@@ -453,7 +453,9 @@ namespace hyperliquid
     {
         std::string vaultAddress;
         bool isDeposit;
-        uint64_t usd;
+        // Whole-and-fractional USD, e.g. 5.0 for $5 - the wire format wants this scaled to
+        // raw USDC units (x1e6), which ExchangeRequestBuilder::vaultTransfer handles.
+        double usd;
     };
 
     struct UsdClassTransferRequest
